[FIX] spreadsheet: faulty logic in filters dependencies
The global filters had to be aware of every type of datasources that it
affected (i.e. pivots, lists and charts). This commit introduces the
possibility for any datasource to register its type in an object handled
by the filters, making the latter agnostic of the different possible
kinds.

Task 2992752

X-original-commit: f15be1e70dcb980074c7141a5b73ea0f98320c7d
Part-of: odoo/enterprise#32526