[IMP] website_sale_stock_renting: rental - search bar
This commit adds more ways for an admin to setup lookup by rental
availabilities.

A new snippet has been added which can be put anywhere on the website to
do a search by date on products.

New options have been added in the customize menu of /shop to add the
rental search bar to the left side menu as well as the new off canvas
options for mobile searches.

TaskId-2806828

closes odoo/enterprise#29909

Related: odoo/odoo#96995
Signed-off-by: William Braeckman (wbr) <wbr@odoo.com>
Co-authored-by: Alexandre De Pape <alpa@odoo.com>
Co-authored-by: William Braeckman <wbr@odoo.com>