[FIX] account_reports: fix unfold on section lines
Until now, if a section line in a report had the attribute
`foldable` set to `True`, the unfolding was not done properly.
Only the first children were unfolded. But these children often
had children of their own, and they remained invisible.

This PR fixes the issue.

At the same time, we also make sure that the `unfoldable` key
in the static line's dict only carries a boolean instead of the
records themselves (which was useless).

task-2971538

closes odoo/enterprise#31458

Signed-off-by: Olivier Colson (oco) <oco@odoo.com>