[FIX] sale_subscription: avoid to recompute the price_unit child upsell/renew
lines

Before this commit, when a renew/upsell line has a parent, the price_unit
was recomputed. If a custom value was set, it was not propagated on the child.

This commit prevent the value to be changed recompuited. The price_unit
can still be changed by hand if needed but we think overriding the price_unit
will have more drawbacks for the user than keeping the parent value.

Taskid: 2988410
X-original-commit: f8d24352a608651975b3c09908b33584cec048dc
Part-of: odoo/enterprise#32173