[FIX] sale_subscription: avoid start_date > next_invoice_date
Before this commit, in some cases, the start_date of a subscription
could be larger than the next_invoice_date if the order was not confirmed
the day of the creation. It would trigger issues. This commit make sure the
ext_invoice-date is not set to its default value when the sale order is not
in a relevant state (confirmed).

Taskid: 2988410
Part-of: odoo/enterprise#31580