[FIX] account_sepa: new pain version
The previous SEPA version used for Germany will be obsolete in November. Germany now uses 001.001.03 instead.

The technical label of the corresponding value of res.company's pain_version field hasn't been changed in order not to break the stability policy, but will be modified in master.

source: [LINK]
task-id: 2985052

[LINK]: https://www.ebics.de/de/datenformate/format-lifecycle

X-original-commit: de53d797237c75ca9d1a19fc2768106609944f19
Part-of: odoo/enterprise#33474