[REF] planning: convert form to wowl
This commit converts the form view to wowl.
Also, it corrects the view.

Note :
The form view is called from gantt, which is still in legacy.
As a consequence, this commit just aims to make the converted view
available when it will be needed, but the legacy view remains in use.

task-2922976

closes odoo/enterprise#31484

Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>