[REF] planning: convert form to wowl
This commit converts the form view to wowl.
Also, it corrects the view.

Note :
The form view is called from gantt, which is still in legacy.
As a consequence, this commit just aims to make the converted view
available when it will be needed, but the legacy view remains in use.

task-2922976

This commit is a backport of 6ff687bf08ee7d9d2f80673df5d848f835eeea8f

Part-of: odoo/enterprise#32058