[FIX] sale_subscription: duplicate correctly upsells
Current behaviour:
When duplicating an upsell, the new duplicated upsell is not linked to the initial subscription.
(It doesn't appear in the "history" view (click the "Sales" smart button on the subscription),
and there is no smart button on the new upsell linking back to the subscription)

Expected behaviour:
A duplicated upsell should be linked to the subscription,
with the smart button allowing for easy back and forth between the upsell and the subscription

Steps to reproduce:
- Create an upsell to a subscription
- Duplicate the upsell SO
- Observe that the new SO has no "Sales" smart button, and in the original subscription,
when clicking on the "Sales" smart button, the newly created SO is not present.

Reason for the problem:
After the refactoring of moving the subscription code to sale_order,
there was no override of the copy_data method.
When an upsell was duplicated, the new order client_order_ref was empty,
subscription_id was also empty, and origin_order_id was set to the id of the source upsell SO,
preventing it to be computed by _compute_origin_order_id correctly
(since this compute expect the origin_order_id to be empty to set it to the subscription_id).

Fix:
Explicitly copy the problematic fields (client_order_ref, subscription_id, origin_order_id)
from the source when duplicating.

Affected versions:
- saas-15.3
- master

opw-2983856

closes odoo/enterprise#31686

X-original-commit: c65418c654b51832a508c9e9f56c946a0b5b73b2
Signed-off-by: Grazioso Andrea (agr) <agr@odoo.com>
Signed-off-by: Piryns Victor (pivi) <pivi@odoo.com>