[IMP] l10n_mx_edi_40: heuristic for the legal names
With the introduction of cfdi 4.0, the MX gov decided
that the name given of the supplier and the customer
should be an official name.  (so they would check if
RFC and name match or something)

But sometimes people put the name of their company or
customer as e.g. Odoo SA or Odoo Technologies SL to
include the type of company, but in the XML it will not
want the SA, SL or S de RL (de CV) and as such we
put a little heuristic that removes it automatically
when the company name or partner name has it to hopefully
already match the official name with the MX gov.

closes odoo/enterprise#32054

Task-id: 2985800
X-original-commit: 287b1c92dfbe587a363750c56065af59fd00019d
Signed-off-by: Quentin De Paoli <qdp@odoo.com>
Signed-off-by: Josse Colpaert <jco@odoo.com>