[FIX] account_sepa: new pain version
The previous version of SEPA used for Germany was deprecated in early November. A PR was made to change this in the previous version (see: https://github.com/odoo/enterprise/pull/31470) but to avoid breaking the stable policy, the previous fix did not remove the function or change the data in the database.

This PR removes the unnecessary function, and changes the data to be more accurate with the current version of the pain.

closes odoo/enterprise#31542

Task-id: 3063601
Related: odoo/upgrade#3912
Signed-off-by: John Laterre (jol) <jol@odoo.com>