[FIX] knowledge: remove emojis from xml files
On some macOS systems running some builds of lxml (need further investigation
for specific versions and OS), lxml parser is not able to decode certain emojis.
This could prevent someone to install the knowledge module on such systems.

This fix removes every emojis stored in an xml file so that anyone can install
Knowledge. This commit should be reverted when a better solution is found.

Task-2989117

closes odoo/enterprise#31627

X-original-commit: 56a40f8569b8ebeeab9b38c00bd59b2cac1b597d
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>