[FIX] project_enterprise, industry_fsm: change planned dates visibility
As we want to encourage the use of the gantt for planning tasks, the fields are

removed from the form view.

As project sharing does not display the gantt view, the fields are preserved in

the form view but the visibility is changed so that they are only displayed when

one of them are set.

task-2895909

closes odoo/enterprise#32499

X-original-commit: f5d88884a279ddb170a3e2a795b95100a920a593
Related: odoo/odoo#102633
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>