[FIX] helpdesk: section field need selection
When defining a selection type field in fieldsToFetch, it is necessary
to have a selection. We can use an empty array if it is not used by
our field.

In legacy, a selection field with no selection was supported.
It worked only if the field is read and not edited. For the case of
helpdesk_sla_many2many_tags, it worked correctly because the selection
field is not modified. We only read its value.

Now, we prefer not to support the absence of selection because it forces
the developer to be aware of the possible selections he gives. If he
only wants to read the value of the field and not modify it, he can
pass an empty list.

closes odoo/enterprise#31544

Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>