[FIX] mrp_workorder: translate embed form view in owl
The two form views contains on the workorder tablet view are
rendered in the new Owl framework thanks to this commit.
The FormController is overrided to save the two models (mrp.workorder
and quality.check) each time a click is triggered on a button before the
doing the action of that very button.

Task : 2991906

X-original-commit: 4dc8659777ea1ec3050af9e92e7ff03a816eb551
Part-of: odoo/enterprise#32172