[FIX] sale_ebay: default ebay quantity value for product:
Steps to reproduce:
- Don't save between the steps only at the end
- Create a product
- Check sell on ebay and put fixed price
- Configure some variants
- Save

Current behavior:
The first variant in the ebay tab list has an ebay quantity
of zero

Expected behavior:
The quantity should be one (classical default value)

Explanation:
Ebay_quantity is a related field between the product template
and the product variant. The product template is created after
the product variants so when the product template is created
with a default value for ebay_quantity of zero it also sets
the value of the related product variant.

opw-2974596

closes odoo/enterprise#32258

X-original-commit: ab482b8f9dec2336d46a4f56410bd044c84a9e46
Signed-off-by: Fockedey Martin (mafo) <mafo@odoo.com>