[FIX] account: fix reco widget deleting selected lines
Since v16, selecting a reconciliation model in the new reco widget would undo the previous selection.

That was because of an uncomplete check at the creation of the reco model lines.

Took the opportunity to fix a few display issues and some forgotten indentation error who just appeared mysteriously.

task-2988109

closes odoo/enterprise#31579

Signed-off-by: Laurent Smet <las@odoo.com>