[FIX] l10n_be_hr_payroll_*: fix holiday pay recovery
Whenever an employee has a simple holiday pay to recover, the amount to recover is
only depending on the number of days and is not capped.

If the employee earns less in his/her new company, the holiday pay amount to recover
should be capped to the employee cost for the number of days to recover.

Example:
If an employee has 3000€ of simple holiday pay, 10 days to recover and costs
250€ per day in his/her previous company.

Before this commit, the total amount to be recovered is divided by the number of days
to recover. Meaning that 300€ will be taken from the holiday pay each day the employee is on leave.

After this commit, the amount deducted from the simple holiday pay equals the number of days taken
multiplied by the daily cost of the employee.
Since the employee costs 250€/day, the maximum amount deducted from the simple holiday pay is 2500€
after the 10 days of leave are taken.

task-2984607

closes odoo/enterprise#33469

X-original-commit: fab2d58331f03c0462767d389c703218bd2e1369
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>
Signed-off-by: phwa-odoo <phwa@odoo.com>