[FIX] l10n_lu_reports_electronic: fix BS & PnL reports
Some fields are overridden to have 0,00 amount in financial reports.

In l10n_lu reports:
- Create journal items for the previous year that would appear on either BS or PnL
- Export PnL through "Export ECDF Declaration"
- Some report fields related to the previous year's data have their amount set to 0,00 even though there should be a non-zero value

That is because `_report_useful_fields` checks if report values contain the `parent_code`.
If the `parent_code` is not in the report, it adds the `parent_field` in report values with 0 amount.
But for the previous year `parent_code` != `parent_field`, as `parent_field` is `parent_code + 1`.
Thus, report values would never contain the parent code and the amount of the parent field is overridden to 0.

opw-2989056

closes odoo/enterprise#31750

X-original-commit: 50c143e0469ca0f04d65620b0bd95128cf27f0eb
Signed-off-by: Josse Colpaert <jco@odoo.com>
Signed-off-by: Aliya Tastemirova (alta) <alta@odoo.com>