[FIX] account_reports: fix report exports
In this commit https://github.com/odoo/enterprise/commit/03d9353d85ed80c6e44fd231e5340d2378af8c81,
a new abstract model is introduced to handle custom report's specific logic.

Many methods were moved from 'account.report' to the new handler,
and the exports was left out, so that the method was always
called on the report, even if it sometimes is defined on the custom handler.

This commit fixes the issue.

Part-of: odoo/enterprise#31586