[IMP] knowledge: introduce embedded views
With this commit, users are able to embed views directly within the Knowledge
editor, directly in the body of articles. Those views are completely dynamic and
integrated as any Odoo view used in the web client context: actions can be
launched, records moved or created, ...

This new feature allows notably to insert reports or graphs directly into
articles. It also allows linking views used in daily workflows (e.g. lead
pipeline, tasks kanban view) to explain or highlight part of a process, to
generate reporting / followup articles, ...

To insert an embedded view, users have to:

  Open the view they want to insert;

  Click on the 'Favorite' dropdown menu under the search bar of the control
  panel;

  Click on the 'Insert view in article' menu item;

A modal asks the user to select the article in which the current view should be
inserted. Once the user selects an article, they are redirected to the related
article with the view being appended at the end of the article. Users are then
able to interact with the embedded view as they normally do with standard views,
such as adding or removing filters, opening records, ...

When a view is being inserted, the system takes a snapshot of the active search
filters and uses them by default for the embedded view. This mechanism allows
the users to fully customize the embedded view by filtering out some records or
by grouping the records as they want.

Technical comments:

For performance reasons, the embedded views are lazily loaded when they become
visible on the user screen. This mechanism will drastically reduce the load
spikes when opening an article containing several embedded views. Indeed,
loading a view can be expensive especially when there are many computed fields
to evaluate on many records. We should then be really cautious and only load the
embedded views when it is strictly necessary.

Limits:

Currently, the act_window used to load the embedded view has to be stored in db.
The system does not support action generated by the server. In the future, it
should be possible to support those actions given that it is possible to
retrieve the original action used to open a given view.

Task-2796156

X-original-commit: 0ffc548c0afa7434f40cccf91df62e5e5172e374
Part-of: odoo/enterprise#32129
Co-authored-by: Julien Banken <jbn@odoo.com>
Co-authored-by: Damien Abeloos <abd@odoo.com>