[FIX] web_studio: fix error when editing Priority widget
Current behaviour:
In Studio, when editing one of the text values of a newly added priority widget,
we get a stacktrace. Visible only in debug=0 mode.

Expected behaviour:
You should be able to edit text values of priority widget without error.

Steps to reproduce:
- Install Project and Studio
- Go to Project, choose whatever project
- Select an item and activate Studio
- Add a Priority Field to the form view
- Select the Priority widget and select "Edit values" in the property tab
- Edit one of the text values and try to confirm
- Stacktrace should show up about a TypeError in js

Reason for the problem:
In debug=0 mode, we are retrieving the value of the `<li>` item with jQuery's `data()` method.
The problem is that data() will coerce strings that have a number in them to a js number.
So if value inside the HTML is '2', val in js is 2, not '2'.
This breaks the conditional search of the index due to the strict equality (===)
which doesn't coerce types implicitly.
index is therefor -1 (non of the element satisfy the search),
so on the next line when we try to access `this.selection[index]`, it is `undefined`,
triggering the TypeError when accessing an attribute on the object.

Fix:
Replace jQuery's `data()` with `dataset.value`, which always returns a string (the raw value of value).
`dataset.value` is already used for the same code when debug mode is activated for either simple debug or asset debug
(same file, line 196).

Affected versions:
- 14.0
- 15.0
- saas-15.2
- saas-15.3
- 16.0
- master

opw-2966783

closes odoo/enterprise#31783

X-original-commit: a6680c6201b3feda8bf79cab89a9ee671ca25c2d
Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>
Signed-off-by: Piryns Victor (pivi) <pivi@odoo.com>