[FIX] acount_reports: fix xml save export
Steps to reproduce:
- Install l10n_be_intrastat.
- Open the Accounting App.
- Go to Reporting > Intrastat Report.
- Click on SAVE.
- Select XML as export format > EXPORT.
- Should raise a traceback.

This is due to the fact that the file content was
automatically encoded. But it should not always be the case.

With XMLs for instance, the content is of type `markupsafe.Markup`
and should be kept as is, instead of being encoded.

With this PR, a check is made before encoding the file content,
to make sure it is of type `bytes`.

closes odoo/enterprise#31648

X-original-commit: 812c2dd216aadf7da57e461fded36022c55761b1
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: John Laterre (jol) <jol@odoo.com>