[FIX] account_invoice_extract: fix race condition for date boxes
There was a race condition between the handlers of the blur and click
events for date boxes.
If the blur is handled first, the boxes are removed from the DOM and the
click fails as it happens on an element which isn't present.

For the other field's boxes, there was no issue as the `mousedown` event
was prevented and the field wasn't losing focus, letting the boxes
mounted.
For the date fields, the `mousedown` event wasn't prevented as it was
used to hide the calendar dropdown for UI feedback.

In this new version, the `mousedown` event is always prevented and
a fake <kbd>ESC</kbd> keypress is triggered for date fields to hide
the calendar.
This also has the benefit that the boxes are still displayed after a
click which is better for the user experience.

closes odoo/enterprise#31853

Signed-off-by: Georis François (fge) <fge@odoo.com>