[IMP] knowledge: introduce embedded view links
This commit adds a new block for the view links. A view link is a pseudo-link
that redirects the user to a view. It can be useful to quickly redirect users
somewhere in Odoo.

To create a view link, users have to:

  Open the view they want the link to redirect to;

  Click on the 'Favorite' dropdown menu under the search bar of the control panel;

  Click on the 'Insert link in article' menu item;

A modal asks the user to select an article in which the view link will be
inserted. Once the user selects an article, they are redirected to the selected
article and the view link is appended at the end of the article.

As for the embedded views, the user can pick active search filters and those
filters are set by default when the user opens the link.

Limits:

Currently, the act_window used to load the embedded view has to be stored in db.
The system does not support action generated by the server. In the future, it
should be possible to support those actions given that it is possible to
retrieve the original action used to open a given view.

Task-2796156

Part-of: odoo/enterprise#29423
Co-authored-by: Julien Banken <jbn@odoo.com>
Co-authored-by: Damien Abeloos <abd@odoo.com>