[FIX] {website_}appointment: correct new front end design
Various corrections:

- In validation screen: make the bar above buttons as wide as
the screen. Make outlook and google calendar buttons closer.
- In demo data: The centered text for "Technical Demo" is not
looking good at all. Make it align left.
- In various screens, have the same spacing between the arrow
and the All appointment text in the navbar.
- In the appointment card view, make the horizontal bar as wide
as the rest of the elements.
- Make the titles appear in the same place of the screen to make
transitions smoother. (No "flicker" because of moving element)
- Make navbar the same height in all screens.
- Make user picture appear at the same place in registration and
validation views. (Most elements will now be set at the same place,
with the same style, if possible. This prevents flickering)
- User function is made a bit darker.
- Fa icons spacing is made uniform and selectors simplified
- Align items at the bottom of appointment cards and between each
other. Switch location and Staff since location always appear, it
is easier to align at the very bottom.
- Make vertical spaces even in appointment cards.

User card

Cards are now displayed even when no avatars is set to the
appointment type (except for random assignment, since no user is set
at the start of the appointment process): it now contains the phone
and email of the selected user as well. The texts "Meeting details"
and "You will meet with" are deleted. The templates are simplified
accordingly.
Icons of the user card and of appointment details are aligned in the
"Details" screen.

Task-2991577

closes odoo/enterprise#31655

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>