[IMP] knowledge: owl html_field compatibility, /file, /template
This commit introduces a new system to integrate owl components inside the
Knowledge editor. It will gradually be used for all commands of Knowledge and
will replace the old mechanism. These changes are motivated by the migration of
legacy widgets to owl components.

This commit also migrates the ``/file`` and ``/template`` commands to the new
system. These two commands work as they used to:
  ``/file`` allows users to embed a file into the editor.
  ``/template`` allows users to create a new template into the editor.

No functional change should occur with this commit as it is mainly a code
migration.

Technical comments:

The new Behavior system will do the following actions:

  1) Trigger a command from the wysiwyg widget and render a QWEB template which
  consists of an empty HTMLElement (the `anchor` for a Behavior). It may contain
  a new specific attribute `data-behavior-props` which contains a JSON string
  with the props of the Behavior that needs to be instantiated.

  2) Then, trigger an event for the html_field component in order to create the
  Behavior instance. Since it has to be mounted after the view is already
  rendered, a new OWL App is created, sharing the `env` with the Odoo App. The
  App is mounted in the `anchor` node created by the wysiwyg.

  3) Nodes that need to be saved in the html_field itself are marked with a new
  specific attribute: `data-prop-name` in the Behavior Component template. It is
  parsed when loading an article to create the Behavior instance, being merged
  with the props from `data-behavior-props`.

Task-2796156

X-original-commit: ac9840f318e23eacbe635c1e0323cc5a19c00ae3
Part-of: odoo/enterprise#32129
Co-authored-by: Julien Banken <jbn@odoo.com>
Co-authored-by: Damien Abeloos <abd@odoo.com>