[FIX] sign,sign_itsme: fix HTTP tests failing
In sign the HTTP tests that called the route `/sign/sign` and tested that it is
possible to sign with and without IAP credits when SMS authenticatin is
required were passing the wrong id to the route. Since sign.request and
sign.request.item have the same id when there is only one sign.request.item in
the sign.request, the test was passing in some cases, but not all. This commit
fixes this by passing the correct id to the `sign/sign` route. The same is done
for sign_itsme.

This commit also fixes the inheritance from TestSignController in sign_itsme
that was causing the sign tests to be run again when running the sign_itsme
tests.

closes odoo/enterprise#31687

X-original-commit: 4ca80d05b8a801d1c727a1ab62364b435c571900
Signed-off-by: Xavier Dollé (xdo) <xdo@odoo.com>