[FIX] account_accountant: statement_id in bank reco widget list view
The list view of bank statement line in the bank reco widget does
not display the statement id in the list view.
This is something that is not wanted at this time, as we want to
be able to edit the statement of a bank transaction directly from
there.

Also remove some duplicated records in the xml file.

Finally, add a padding between the partner_id and the statement
in the kanban view.

closes odoo/enterprise#31664

Signed-off-by: Cedric Snauwaert <csn@odoo.com>