[FIX] account_accountant: statement_id in bank reco widget list view
The list view of bank statement line in the bank reco widget does
not display the statement id in the list view.
This is something that is not wanted at this time, as we want to
be able to edit the statement of a bank transaction directly from
there.

Also remove some duplicated records in the xml file.

Finally, add a padding between the partner_id and the statement
in the kanban view.

closes odoo/enterprise#31693

X-original-commit: 1d2fcea61dabb32a38af264b2b9c5e56b3776eef
Signed-off-by: Cedric Snauwaert <csn@odoo.com>
Signed-off-by: Nicolas Viseur <vin@odoo.com>