[FIX] hr_appraisal(_skills): fix ir rule
In hr_appraisal, when a user is added in the manager_ids field and is not the manager of the
employee having his appraisal, he is not able to edit the appraisal.

This commit updates the ir rule to allow user added in the manager_ids field to write the appraisal.

task-2993820

closes odoo/enterprise#31978

X-original-commit: 4d2605a1b3ebc3b055fbdd505f31de68a61f1e21
Signed-off-by: Kevin Baptiste <kba@odoo.com>