[FIX] account_reports: fix totals below sections
Until now, if a line in a report was not foldable, it would not
have a total line, even if the option was activated in the settings.

With this fix, a line that is not foldable but has children,
will also get a total line, as it should.

X-original-commit: e351f7f4eace93b73c8ab64238f5cd7936a920ee
Part-of: odoo/enterprise#32525