[FIX] account_reports: fix totals below sections
Until now, if a line in a report was not foldable, it would not
have a total line, even if the option was activated in the settings.

With this fix, a line that is not foldable but has children,
will also get a total line, as it should.

Part-of: odoo/enterprise#31586