[FIX] web_studio: prevent saving a selection widget with no options
Current behaviour:
In Studio, when removing all options available from the priority widget and saving,
you get a stacktrace, and subsequent access to the view also triggers a stacktrace.

Expected behaviour:
You shouldn't be able to save a selection widget with no default/fallback option.

Steps to reproduce:
- Install Project and Studio
- Go to Project, choose whatever project
- Select an item and activate Studio
- Add a Priority Field to the form view
- Select the Priority widget and select "Edit values" in the property tab
- Remove all options from the widget and save
- You will get a stacktrace and also everytime you try to access the view
you added the widget to (in our case all tasks)

Reason for the problem:
If you save a selection widget with no options, no child `<li>` are appended,
and when we try to retrieve them later on with a `closest('li')` and index it at `[0]`,
it triggers a TypeError, since no `<li>` were found.

Fix:
Show a warning when trying to save a selection field with no options at all.
This behaviour is similar to the already present warning messages when trying
to save relational fields that have no relation specified.

Affected versions:
- 14.0
- 15.0
- saas-15.2
- saas-15.3
- 16.0
- master

opw-2993954

closes odoo/enterprise#31941

X-original-commit: 4676dd9d79f957a0fcfb8922b1303c99e760d442
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>
Signed-off-by: Piryns Victor (pivi) <pivi@odoo.com>