[FIX] delivery_dhl,delivery_fedex,delivery_ups: remove assigned in test
Before this commit, the test was faulty:

```
Traceback (most recent call last):
  File "/home/odoo/src/enterprise/delivery_fedex/tests/test_delivery_fedex.py", line 295, in test_04_fedex_international_delivery_from_delivery_order
    super().test_04_fedex_international_delivery_from_delivery_order()
  File "/home/odoo/src/enterprise/delivery_fedex/tests/test_delivery_fedex.py", line 242, in test_04_fedex_international_delivery_from_delivery_order
    self.assertEqual(delivery_order.state, 'confirmed', 'Shipment state should be waiting(confirmed).')
AssertionError: 'assigned' != 'confirmed'
- assigned
+ confirmed
 : Shipment state should be waiting(confirmed).
```

With the change introduced in https://github.com/odoo/odoo/commit/c7f53f770c1d22fd9aff1823a04c86464a91bf71 ,
the default behavior is to assign pickings right when they are confirmed.

To fix the tests, the `reservation_method` of the stock picking could be
manual, but the default behavior is `at_confirm`, so it makes sense to
just remove the extra step in the tests.

Part-of: odoo/enterprise#33403