[FIX] hr_work_entry_contract_enterprise: hide Payroll icon if no access to hr_contract
Currently, the display of the Payroll account when hr_work_entry_contract_enterprise is installed is not restricted.
A user with no access to hr_contract could see the Payroll icon while not being able to access the app due to an access error on hr_contract.

This commit hides the Payroll app icon if the user doesn't have hr_contract rights.

task-2994285

closes odoo/enterprise#31997

X-original-commit: 09557b9fd5ab041117cda8b0496d0174006be873
Signed-off-by: Kevin Baptiste <kba@odoo.com>
Signed-off-by: phwa-odoo <phwa@odoo.com>