[FIX] web_gantt: prevent start/end_datetime changes when resizing pills
Prior to the fix and in some cases, pill resizing in the gantt view
would lead to unwanted changes in the start/end times of the resized
pills. This behaviour can be seen planning and hr work entries.

This issue stems from a rounding error that is sometimes made when
computing the difference of the pill before and after resizing. The
change in this commit addresses this error by restricting the
difference to be either 0 or any of the cell precisions for the given
scale

task-2996205

closes odoo/enterprise#32537

X-original-commit: 745fa64693138c8fd19120931d5e6ff6334874eb
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>