[FIX] web_gantt: prevent start/end_datetime changes when resizing pills
Prior to the fix and in some cases, pill resizing in the gantt view
would lead to unwanted changes in the start/end times of the resized
pills. This behaviour can be seen planning and hr work entries.

This issue stems from a rounding error that is sometimes made when
computing the difference of the pill before and after resizing. The
change in this commit addresses this error by restricting the
difference to be either 0 or any of the cell precisions for the given
scale

task-2996205

closes odoo/enterprise#32576

X-original-commit: 5673c0e44ef2e1d36eb81e1e7a0cbcde17b1c6a8
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>