[FIX] helpdesk: filter by co and correct exceeded hours calculation
1) `exceeded_hours` field
Prior to this fix, 'Working Hours to SLA Deadline' would only be
calculated if SLA had been reached. This meant its value was 0.0 in db.
The issue comes when this 0.0 is taken into account
for the calculation of the average.
We rather have the number of hours left to reach the SLA deadline.

2) limit report to allowed companies
Prior to this fix, tickets from all companies were taken displayed in
the report. Now, we limit it to the tickets from the allowed companies.

Note: deleted two ir.rules.
\- helpdesk_sla_report_analysis_rule_manager because it denied access to
ticket with a different company from the user's one.
\- helpdesk_ticket_report_analysis_rule_multi_company, because it is no
more than a domain, so we moved it in the action.

task-2994097

closes odoo/enterprise#32780

X-original-commit: 7b6424f0cd9e14438dedcf814be03929bcb60e22
Signed-off-by: Christophe Simonis <chs@odoo.com>
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>