[FIX] account_intrastat: product origin country name translation
Due to the recent jsonb fields update, the product origin country name
as retrieved by the sql intrastat query is in the form of a json string
e.g. {'en_US': 'Algeria'}

To correct this we use the ->> operator to parse the json in the field
and retrieve the correctly translated value for the country name.

closes odoo/enterprise#31871

X-original-commit: 5f7ccf9509067252a476a85b614755832bbb59ac
Signed-off-by: Josse Colpaert <jco@odoo.com>
Signed-off-by: Daniel Kosky (dako) <dako@odoo.com>