[FIX] account_intrastat: product origin country name translation
Due to the recent jsonb fields update, the product origin country name
as retrieved by the sql intrastat query is in the form of a json string
e.g. {'en_US': 'Algeria'}

To correct this we use the ->> operator to parse the json in the field
and retrieve the correctly translated value for the country name.

closes odoo/enterprise#31830

Signed-off-by: Josse Colpaert <jco@odoo.com>