[IMP] payment_sepa_direct_debit: uniformize the name, summary and sequence
Before this commit, most of payment modules had a custom sequence
number intended to sort them in the Apps' kanban view. In reality, the
sort on the module name makes the custom sequence useless. This commit
thus sets all of payment modules' sequences to `350`.

In an effort for uniformization, we also made names and summaries more
generic, and removed the descriptions which did not add any value.

Task - 2960976

closes odoo/enterprise#31845

Related: odoo/odoo#101276
Signed-off-by: Antoine Vandevenne (anv) <anv@odoo.com>