[REF] mrp_workorder_iot: port pedal_form to wowl
The `Tablet` component hosts two forms (mrp.workorder and quality.check).
Instead of implementing the feature of pedal_form to one of the two
forms (previously in mrp.workorder), we are now implementing it on top
of the `Tablet` component.

For future reference, we note the following:

- IoT triggers are defined in the work centers.
- When the `Tablet` component is rendered, we instantiate the device
controllers and start listening before the willStart hook ends.
- During the operational time of the `Tablet` component, when a
trigger is pressed, a corresponding barcode event, prefixed with
`O-BTN.`, will be triggered and the barcode service will handle
the event by automatically clicking the button with appropriate
barcode_trigger.
- When leaving the `Tablet` component, the listeners are stopped.
- The status button is put on top of the step items and is hidden
if there are no iot triggers.

closes odoo/enterprise#31883

X-original-commit: e131ca1f7a7e49833e248ed9bc0dfe60ab11d860
Signed-off-by: William Henrotin (whe) <whe@odoo.com>
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>