[IMP] account_edi_ubl_cii: always generate Factur-X silently
Always generate a Factur-X xml and embed it when a PDF is created. But no edi.document
is created, no warnings will be displayed on the invoice, and the PDF is not PDF/A-3.
The only purpose here is the Odoo interoperability.

Rename the 'Factur-X' option to 'Factur-X PDF/A-3' and make it only visible for FR and DE companies.

This 'Factur-X PDF/A-3' option will trigger the regular account.edi flow: create an edi.document
and show warnings on the move if errors are detected. When a PDF is printed, embed the Factur-X xml
and make the PDF a PDF/A-3. Thus, it can be used on Chorus Pro.

task-2957808

closes odoo/enterprise#32344

X-original-commit: e52ab3a24d963045336bcc0871437b234efe97d9
Related: odoo/odoo#102311
Related: odoo/documentation#2789
Signed-off-by: Laurent Smet <las@odoo.com>
Signed-off-by: Julien Van Roy <juvr@odoo.com>