[FIX] voip: adapt code to PhoneField changes
This commit is the counterpart of https://github.com/odoo/odoo/pull/101433,
and  adapts the code of voip to patch the right elements in the
PhoneField template. As the field now always displays a Call button,
the call must be enabled even in edition mode.

Tests have been adapted with a new selector

closes odoo/enterprise#32108

X-original-commit: 944ec9f4a77376d8f21af594e8fe4020868f5fa0
Related: odoo/odoo#101786
Signed-off-by: Michaël Mattiello <mcm@odoo.com>
Signed-off-by: Luca Vitali <luvi@odoo.com>