[FIX] voip: adapt code to PhoneField changes
This commit is the counterpart of https://github.com/odoo/odoo/pull/101433,
and  adapts the code of voip to patch the right elements in the
PhoneField template. As the field now always displays a Call button,
the call must be enabled even in edition mode.

Tests have been adapted with a new selector

closes odoo/enterprise#31954

Related: odoo/odoo#101433
Signed-off-by: Michaël Mattiello <mcm@odoo.com>