[FIX] stock_barcode_product_expiry: tour date
Before this commit, the dates checked in the test tour
`test_gs1_receipt_expiration_date` were hard written. But as the tours
are runned by chrome, the way the dates are generated is depending of
chrome's language setting and for some OS, the chrome's language is the
OS one and is not configurable.
So, this tour will failed if run on a machine with a different date
format (which depends of the language).
To fix that, this commit generates the date in the test.

Part-of: odoo/enterprise#30979