[REM] project_timesheet_forecast(_contract,_sale): revert "generate timesheet in planning" tasks
This commit reverts two tasks that introduced the "Generate timesheet"
action in the Planning app. Prior to this commit, the user could select
a specific resource and click on the "generate timesheet" button in order
to generate timesheets for that resource-task-shift combination.

This commit removes this functionality. This should make the planning app
simpler and easier to use.

task-2999234

closes odoo/enterprise#31996

Related: odoo/upgrade#3934
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>