[FIX] sale_subscription: qty_invoiced was not properly computed for refund
Before this commit, the qty_invoiced of subscription added all the related invoices.
Refund were added to the qty_invoiced. See this example: a subscription with 2 lines
P1, Qty 1
P2, Qty 2

If we invoice both lines, qty invoiced will be equal to Qty 1 and Qty 2
If we refund the invoice, the expected qty_invoiced value is 0 for both lines.

closes odoo/enterprise#31908

closes odoo/enterprise#32347

X-original-commit: d274ee5b3ea1f8db66cfe871b4cbd913f8abd8ff
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>
Signed-off-by: Arnaud Joset <arj@odoo.com>