[FIX] web_grid: use newest groupBy values
Before this PR it is not possible to override the read_grid() and insert multiple custom keys for the titles in the Grid view.
Since Odoo was not re-checking/updating the groupBy part it was always loading the old values.
This caused it to be impossible to define custom row_fields (with a super()) in the read_group.

By always updating/computing/comparing against what is controller in the XML-RPC layer we always have all the row_fields used and shown.

Sample to test:
In timesheet_grid/models/analytic.py in the 'read_grid' set a custom row_fields right above the if not self.env.context.get() like this:
row_fields = ['partner_id', 'project_id', 'task_id', 'so_line']
This should show the customer name, projct name, task name and sale order name (if set) in the Grid view.
Before our JS changes it doesn't though as it keeps working with the defaults ('project_id', 'task_id') and doesn't use our custom set ones.

closes odoo/enterprise#32060

X-original-commit: fa11970519f8f04ffa2f210f4262bb41465d6b06
Signed-off-by: Géry Debongnie <ged@odoo.com>