[FIX] sign: restore sign request activity
Before this PR, the sign request activity was not converted to wowl.
Moreover, a traceback occured when executing the sign request action.

The traceback occured because we were trying to execute the
`sign.SignableDocument` action even if the user had no authority to
sign the document.

This PR restores the sign request activity and fixes this issue.

task-2998137

closes odoo/enterprise#31952

Related: odoo/odoo#101469
Signed-off-by: Sébastien Theys (seb) <seb@odoo.com>