[FIX] knowledge: fix no content display in form view
Purpose:
--------

Instead of opening other users' private articles that are not shown in the
sidebar (but that you can access because you are admin user) when there
are no articles in the sidebar, show the No Content Helper.

When the No Content Helper is shown, the "virtual article in creation" is
now not editable at all (previously, even with the helper displayed, one
could write in the body, change the name, add an icon/cover or add to
favorites). We force the user to use one of our "create" buttons to create
an article.

Task-2985591

closes odoo/enterprise#31947

closes odoo/enterprise#32109

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>