[FIX] sign: fix sign all promise being resolved before necessary
After https://github.com/odoo/enterprise/pull/30505, a new promise for resizing
the frame image was added. However, it was using

```
promise.then(() => {
    promise.then()
})
```
which was causing the second then to be executed afterwards as the nested
promise was not being returned. Functionally this caused the check for
completion of the document to be ran before the data is correctly set. So, the
validate and sign button wasn't appearing.

Also, since the async work doesn't have to be
sequential, it's possible to optimize this by using

```
Promise.all([promise1, promise2]).then(...)
```

task-3000534

closes odoo/enterprise#33858

X-original-commit: a7cfc8b219d4ae00572402f4c844b7b537f451b9
Signed-off-by: Arnaud Joset <arj@odoo.com>