[FIX] sale_subscription: fix alerts configuration display
The Alerts menu item is now hidden for users who are not allowed to access it.
Before this commit, it was visible for some users (such as 'demo'), and clicking it produced the following warning:

odoo odoo.http: You are not allowed to access 'Automated Action' (base.automation) records.
This operation is allowed for the following groups:
    - Administration/Settings
Contact your administrator to request access if necessary.
Implicitly accessed through 'Automated Action' (base.automation).

and the following client traceback:

OwlError: An error occured in the owl lifecycle (see this Error's "cause" property)
    at handleError (http://localhost:8069/web/assets/430-200dc52/web.assets_common.min.js:1238:101)
    at owl.App.handleError (http://localhost:8069/web/assets/430-200dc52/web.assets_common.min.js:2043:29)
    at ComponentNode.initiateRender (http://localhost:8069/web/assets/430-200dc52/web.assets_common.min.js:1537:19)
Caused by: RPC_ERROR: Odoo Server Error
    at makeErrorFromResponse (http://localhost:8069/web/assets/431-cf63903/web.assets_backend.min.js:951:163)
    at XMLHttpRequest.<anonymous> (http://localhost:8069/web/assets/431-cf63903/web.assets_backend.min.js:958:13)

task-3001169

closes odoo/enterprise#32111

X-original-commit: d1d62a41a3c12ac628dc945e365d2092a77f6fc9
Signed-off-by: Arnaud Joset <arj@odoo.com>